Update ATL07 point cloud classifier notebook (#29)
* Rename photon to point cloud

ATL07 points are not actually photons, but an aggregate of ATL03 points.

* Show inference results in confusion matrix and closing words

Demonstrate how the trained model can be used to produce classification results in a new column of the GeoDataFrame, and compare those predictions with the groundtruth in a confusion matrix plot. Added some closing words about data-centric and model-centric ways of improving the results, and added the Petty et al. 2021 paper to the citation list as credit to Alek's help.